+++ source/main/ipc/index.js
@@ -7,6 +7,7 @@ import getLogsApi from './get-logs';
 import resizeWindowApi from './resize-window';
 import loadAsset from './load-asset';
 import getGpuStatus from './get-gpu-status';
+import getRecoveryWalletIdChannel from './getRecoveryWalletIdChannel';
 import { handleBugReportRequests } from './bugReportRequestChannel';
 import { handleFileMetaRequests } from './generateFileMetaChannel';
 import { handlePaperWalletRequests } from './generatePaperWalletChannel';
@@ -33,5 +34,6 @@ export default (window: BrowserWindow) => {
   openExternalUrlChannel;
   // eslint-disable-next-line no-unused-expressions
   openLocalDirectoryChannel;
+  getRecoveryWalletIdChannel();
   handleElectronStoreChannel();
 };