+++ source/renderer/app/actions/wallet-settings-actions.js
@@ -19,11 +19,20 @@ export default class WalletSettingsActions {
     newPassword: string,
     isLegacy: boolean,
   }> = new Action();
-  exportToFile: Action<WalletExportToFileParams> = new Action();
-  startWalletUtxoPolling: Action<any> = new Action();
-  stopWalletUtxoPolling: Action<any> = new Action();
   forceWalletResync: Action<{
     walletId: string,
     isLegacy: boolean,
   }> = new Action();
+  exportToFile: Action<WalletExportToFileParams> = new Action();
+
+  /* ----------  UTXO  ---------- */
+  startWalletUtxoPolling: Action<any> = new Action();
+  stopWalletUtxoPolling: Action<any> = new Action();
+
+  /* ----------  Recovery Phrase Verification  ---------- */
+  recoveryPhraseVerificationContinue: Action<any> = new Action();
+  recoveryPhraseVerificationCheck: Action<{
+    recoveryPhrase: Array<string>,
+  }> = new Action();
+  recoveryPhraseVerificationClose: Action<any> = new Action();
 }