+++ source/renderer/app/components/wallet/settings/WalletRecoveryPhraseStep4Dialog.js
@@ -51,23 +51,23 @@ export const messages = defineMessages({
 
 type Props = {
   onClose: Function,
-  onVerifyAgain: Function,
+  onContinue: Function,
   openExternalLink: Function,
 };
 
 @observer
-export default class WalletRecoveryPhraseStep4 extends Component<Props> {
+export default class WalletRecoveryPhraseStep4Dialog extends Component<Props> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
   render() {
     const { intl } = this.context;
-    const { onClose, onVerifyAgain, openExternalLink } = this.props;
+    const { onClose, onContinue, openExternalLink } = this.props;
 
     const actions = [
       {
         label: intl.formatMessage(messages.recoveryPhraseStep4Button),
-        onClick: onVerifyAgain,
+        onClick: onContinue,
         className: 'attention',
       },
     ];