+++ tests/wallets/e2e/steps/helpers.js
@@ -1,6 +1,5 @@
 // @flow
 import { expect } from 'chai';
-import BigNumber from 'bignumber.js/bignumber';
 import { expectTextInSelector, waitAndClick } from '../../../common/e2e/steps/helpers';
 import { rewardsMnemonics, balanceMnemonics, balanceItnMnemonics, testStorageKeys } from '../../../common/e2e/steps/config';
 import { WalletSyncStateStatuses } from '../../../../source/renderer/app/domains/Wallet';
@@ -12,15 +11,6 @@ const ADD_WALLET = '.WalletAdd';
 const IMPORT_WALLET_BUTTON = '.importWalletButton';
 const IMPORT_WALLET_DIALOG = '.WalletFileImportDialog';
 const DEFAULT_LANGUAGE = 'en-US';
-
-export const addOrSetWalletsForScenario = function(wallet: Object) {
-  this.wallet = wallet;
-  if (this.context.wallets != null) {
-    this.context.wallets.push(this.wallet);
-  } else {
-    this.context.wallets = [this.wallet];
-  }
-};
 
 let rewardsMnemonicsIndex = 0;
 export const noWalletsErrorMessage = `The balance wallet for funds transfering was already used and has no longer funds.
@@ -83,13 +73,8 @@ export const restoreLegacyWallet = async (
     recoveryPhrase = null;
   }
   await client.executeAsync((name, recoveryPhrase, transferFunds, noWalletsErrorMessage, done) => {
-    let mnemonics = recoveryPhrase || daedalus.utils.crypto.generateMnemonic(12);
+    const mnemonics = recoveryPhrase || daedalus.utils.crypto.generateMnemonic(12);
     const recoveryPhraseArray = typeof mnemonics === 'string' ? mnemonics.split(' ') : mnemonics;
-    done({
-      walletName: name,
-      recoveryPhrase: recoveryPhraseArray,
-      spendingPassword: 'Secret1234',
-    })
     daedalus.api.ada
       .restoreByronRandomWallet({
         walletName: name,
@@ -137,8 +122,11 @@ export const getNameOfActiveWalletInSidebar = async function() {
   );
 };
 
-export const getWalletByName = function(walletName: string) {
-  return this.context.wallets.find(w => w.name === walletName);
+export const getWalletByName = async function(walletName: string) {
+  const wallet = await this.client.execute(walletName => (
+    daedalus.stores.wallets.getWalletByName(walletName)
+  ), walletName);
+  return wallet.value;
 };
 
 /**
@@ -269,7 +257,7 @@ export const createWallets = async function(
 const createWalletsSequentially = async function(wallets: Array<any>) {
   const isIncentivizedTestnetRequest = await this.client.execute(() => global.isIncentivizedTestnet);
   for (const walletData of wallets) {
-    const result = await this.client.executeAsync((wallet, isIncentivizedTestnet, done) => {
+    await this.client.executeAsync((wallet, isIncentivizedTestnet, done) => {
       const mnemonic = daedalus.utils.crypto.generateMnemonic(isIncentivizedTestnet ? 15 : 12);
       daedalus.api.ada
         .createWallet({
@@ -290,22 +278,19 @@ const createWalletsSequentially = async function(wallets: Array<any>) {
         )
         .catch(error => done(error.stack));
     }, walletData, isIncentivizedTestnetRequest.value);
-    const wallet = await waitUntilWalletIsLoaded.call(this, walletData.name);
-    addOrSetWalletsForScenario.call(this, wallet);
+    await waitUntilWalletIsLoaded.call(this, walletData.name);
   }
 };
 
 const createWalletsAsync = async function(table, isLegacy?: boolean) {
   const result = await this.client.executeAsync((wallets, isLegacyWallet, done) => {
     const mnemonics = {};
-    const { restoreLegacyRequest, walletsRequest } = daedalus.stores.wallets;
     const { restoreByronRandomWallet, createWallet } = daedalus.api.ada;
-    const request = isLegacyWallet ? restoreLegacyRequest : walletsRequest;
     const apiEndpoint = isLegacyWallet ? restoreByronRandomWallet : createWallet;
     const mnemonicsLength = isLegacyWallet ? 12 : 15;
 
     window.Promise.all(
-      wallets.map((wallet, index) => {
+      wallets.map(wallet => {
         const mnemonic = daedalus.utils.crypto.generateMnemonic(mnemonicsLength);
         const recoveryPhrase = !isLegacyWallet
           ? mnemonic
@@ -320,22 +305,15 @@ const createWalletsAsync = async function(table, isLegacy?: boolean) {
         });
       })
     )
-      .then(storeWallets =>
+      .then(() =>
         daedalus.stores.wallets
           .refreshWalletsData()
-          .then(() => done({ storeWallets, mnemonics }))
+          .then(() => done({ mnemonics }))
           .catch(error => done(error))
       )
-      .catch(error => done('error.stack'));
+      .catch(error => done(error));
   }, table, isLegacy);
 
-  const { storeWallets, mnemonics, BLAH } = result.value;
-
-  if (this.context.wallets != null) {
-    this.context.wallets.push(...result.value.storeWallets);
-  } else {
-    this.context.wallets = result.value.storeWallets;
-  }
   this.mnemonics = Object.assign(
     {},
     result.value.mnemonics,